collector_alphavantage_news/requirements.txt
requests==2.32.3
pydantic==2.9.2
google-cloud-pubsub==2.23.0
python-dotenv==1.0.1
